# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Librería API",
        "description": "Documentación de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/direcciones/{id}": {
            "get": {
                "tags": [
                    "direccion-controller"
                ],
                "operationId": "getDireccionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Direccion"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "direccion-controller"
                ],
                "operationId": "updateDireccion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DireccionUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Direccion"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "direccion-controller"
                ],
                "operationId": "deleteDireccion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/comunas/{id}": {
            "get": {
                "tags": [
                    "comuna-controller"
                ],
                "operationId": "getComunaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comuna"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "comuna-controller"
                ],
                "operationId": "updateComuna",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comuna"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comuna"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "comuna-controller"
                ],
                "operationId": "deleteComuna",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/autores/{id}": {
            "get": {
                "tags": [
                    "autor-controller"
                ],
                "operationId": "getAutorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "autor-controller"
                ],
                "operationId": "updateAutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "autor-controller"
                ],
                "operationId": "deleteAutor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/direcciones": {
            "get": {
                "tags": [
                    "direccion-controller"
                ],
                "operationId": "getAllDirecciones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Direccion"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "direccion-controller"
                ],
                "operationId": "createDireccion",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DireccionCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Direccion"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comunas": {
            "get": {
                "tags": [
                    "comuna-controller"
                ],
                "operationId": "getAllComunas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Comuna"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "comuna-controller"
                ],
                "operationId": "createComuna",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comuna"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comuna"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/autores": {
            "get": {
                "tags": [
                    "autor-controller"
                ],
                "operationId": "getAllAutores",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Autor"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "autor-controller"
                ],
                "operationId": "createAutor",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Autor"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Autor"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Autor": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
...
            "DireccionUpdateRequest": {
                "type": "object",
                "properties": {
                    "calle": {
                        "type": "string"
                    },
                    "numero": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "comunaId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "Comuna": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    }
                }
            },
            "Direccion": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "calle": {
                        "type": "string"
                    },
                    "numero": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "comuna": {
                        "$ref": "#/components/schemas/Comuna"
                    }
                }
            },
...
            "DireccionCreateRequest": {
                "type": "object",
                "properties": {
                    "calle": {
                        "type": "string"
                    },
                    "numero": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "comunaId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            }
        }
    }
}