# src/main/java/com/example/libreria_app/dto/DireccionUpdateRequest.java
package com.example.libreria_app.dto;

public class DireccionUpdateRequest {
    private String calle;
    private Integer numero;
    private Long comunaId;

    public DireccionUpdateRequest() {
    }

    public String getCalle() {
        return calle;
    }

    public void setCalle(String calle) {
        this.calle = calle;
    }

    public Integer getNumero() {
        return numero;
    }

    public void setNumero(Integer numero) {
        this.numero = numero;
    }

    public Long getComunaId() {
        return comunaId;
    }

    public void setComunaId(Long comunaId) {
        this.comunaId = comunaId;
    }
}